requirements.txt
flask
statsmodels
tabulate
matplotlib
mysql-connector-python
flask_mysqldb
flask-login
